+++ src/main/java/org/example/biblijava/controller/auth/LoginController.java
@@ -1,5 +1,14 @@
 package org.example.biblijava.controller.auth;
 
+import java.io.IOException;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+import org.example.biblijava.BibliJavaApplication;
+import org.example.biblijava.util.DatabaseUtil;
+
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -8,13 +17,6 @@ import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
-import org.example.biblijava.util.DatabaseUtil;
-
-import java.io.IOException;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 
 public class LoginController {
 
@@ -23,6 +25,12 @@ public class LoginController {
     @FXML
     private PasswordField passwordField;
 
+    private BibliJavaApplication mainApp;
+
+    public void setMainApp(BibliJavaApplication mainApp) {
+        this.mainApp = mainApp;
+    }
+
     @FXML
     private void handleLogin() {
         String username = usernameField.getText();
@@ -30,6 +38,7 @@ public class LoginController {
 
         if (authenticateUser(username, password)) {
             showAlert(Alert.AlertType.INFORMATION, "Login Successful!", "Welcome " + username);
+            mainApp.setAuthenticated(true);  // Mettre à jour l'état d'authentification
             closeLoginWindow();
         } else {
             showAlert(Alert.AlertType.ERROR, "Login Failed", "Incorrect username or password.");
