+++ src/main/java/org/example/biblijava/util/DatabaseUtil.java
